# specs/001-backend-api-data/contracts/tasks.openapi.yaml
openapi: 3.0.3
info:
  title: Todo Backend API
  version: 0.1.0
  description: |
    OpenAPI contract for task operations.

    Authentication:
    - All endpoints require `Authorization: Bearer <token>`.
    - The backend derives the user identity from the validated token and scopes all task
      operations to that user.
servers:
  - url: http://localhost:8000

tags:
  - name: tasks

paths:
  /v1/tasks:
    get:
      tags: [tasks]
      summary: List tasks
      description: Returns tasks owned by the authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - name: completed
          in: query
          required: false
          schema:
            type: boolean
          description: Optional filter by completion state.
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [tasks]
      summary: Create task
      description: Creates a new task owned by the authenticated user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreateRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"

  /v1/tasks/{task_id}:
    get:
      tags: [tasks]
      summary: Get task
      description: Returns a single task owned by the authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [tasks]
      summary: Update task
      description: Updates mutable fields on a task owned by the authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdateRequest"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      tags: [tasks]
      summary: Delete task
      description: Deletes a task owned by the authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /v1/tasks/{task_id}/toggle:
    post:
      tags: [tasks]
      summary: Toggle completion
      description: Toggles completion state for a task owned by the authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
      description: Task identifier.

  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: unauthorized
            message:
              type: string
              example: Missing or invalid token.
            details:
              type: object
              additionalProperties: true

    Task:
      type: object
      required: [id, title, completed, created_at, updated_at]
      properties:
        id:
          type: string
        title:
          type: string
        details:
          type: string
          nullable: true
        completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    TaskResponse:
      type: object
      required: [task]
      properties:
        task:
          $ref: "#/components/schemas/Task"

    TaskListResponse:
      type: object
      required: [tasks]
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"

    TaskCreateRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        details:
          type: string
          maxLength: 2000

    TaskUpdateRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        details:
          type: string
          maxLength: 2000

  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"